requirements.txt
numpy
tqdm
pandas
Pillow
torch
matplotlib
ipython
tensorflow
torchvision